requirements.txt
Flask==2.3.3
Flask-Sock==0.7.0
simple-websocket==0.10.1
pandas
openpyxl
